edited field components to allow page bg colour, moved repairAttemptForm to components, adjusted names to not be conflicting
